feat(diff): add get_diff_content tool, sharing a new UnifiedDiffGenerator with DiffViewContent
Extracts the git-diff/LCS unified diff generation out of DiffViewContent into a
shared, self-contained UnifiedDiffGenerator (own temp-dir lifecycle, ignoreWhitespace
as a parameter) so both the classic diff view and the new get_diff_content MCP tool
compute diffs identically. DiffService caches the last ShowDiff call's inputs (under
a lock, since get_diff_content runs off the UI thread) so get_diff_content can return
a bounded-size unified diff instead of the caller re-reading/re-diffing large files
itself. Also fixes header-line path leakage (raw temp file paths in diff --git/---/+++
lines instead of a/original, b/modified) and a spurious BOM in the first diff line,
both inherited from the original RunGitDiff/File.WriteAllText(..., Encoding.UTF8) code.

Design discussion and independent review notes in _notes/get-diff-content-proposal-issue.md.
