+++ ClarionAssistant/ClarionAssistant.csproj
@@ -188,6 +188,7 @@
     <Compile Include="Terminal\ModernEmbeditorViewContent.cs" />
     <Compile Include="Terminal\MonacoEditorControl.cs" />
     <Compile Include="Services\DiffService.cs" />
+    <Compile Include="Services\UnifiedDiffGenerator.cs" />
     <Compile Include="Services\EncodingHelper.cs" />
     <Compile Include="Services\DocGraphService.cs" />
     <Compile Include="Services\InstanceCoordinationService.cs" />
